# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation",
    "version": "1.0.0",
    "title": "Swagger Proteins",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "carloseduardo.andrade@live.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Proteins",
      "description": "All proteins"
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/proteins?page=1&limit=10": {
      "get": {
        "summary": "List proteins",
        "description": "Return 10 results on page 1",
        "operationId": "listAllProtein",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          "404": {
            "description": "Invalid ID"
          }
        }
      }
    },
    "/proteins/{id}": {
      "get": {
        "summary": "Find protein by ID",
        "description": "Returns a single protein",
        "operationId": "findProteinById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of protein to return",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/Protein"
            }
          },
          "404": {
            "description": "Protein not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Protein": {
      "type": "object",
      "properties": {
        "protein_accession": {
          "type": "string",
          "example": "URS0000000001"
        },
        "description": {
          "type": "string",
          "example": "serine hydroxymethyltransferase"
        },
        "label": {
          "type": "string",
          "example": "glyA"
        },
        "synonyms": {
          "type": "array",
          "example": "A0123_00153",
          "items": {
            "type": "string"
          }
        }
      },
      "xml": {
        "name": "Protein"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}